Added tests for case 79.
It takes a little bit of work to enable testing. caseguard.py must be moved in
a Mercurial source folder, under hgext/, and the test files in tests/

The patch must be applied to tests/hghave. Build Mercurial and run the tests:

make test-caseguard*

Depending on the operating system you're using, you might run two or three of
the tests (one test is dependent on having a case-sensitive filesystem)